+++ indico/modules/events/registration/controllers/display.py
@@ -125,17 +125,17 @@ class RHParticipantList(RHRegistrationFormDisplayBase):
     def _is_checkin_visible(reg):
         return reg.registration_form.publish_checkin_enabled and reg.checked_in
 
-    def _picture_url(self, reg, field_data_id=None):
-        if not field_data_id:  # by default, return url for main picture
-            field_data_id = reg.picture_data.field_data_id if reg.picture_data else None
-        return url_for('event_registration.participant_picture', reg, field_data_id=field_data_id) \
-            if field_data_id else ''
+    def _get_picture_url(self, reg, field_data):
+        if not field_data or field_data.storage_file_id is None:
+            return ''
+        return url_for('event_registration.participant_picture', reg, field_data_id=field_data.field_data_id)
 
     def _merged_participant_list_table(self):
         def _process_registration(reg, column_names):
             personal_data = reg.get_personal_data()
             columns = [{'text': personal_data.get(column_name, '')}
-                       if column_name != 'picture' else {'text': self._picture_url(reg), 'is_picture': True}
+                       if column_name != 'picture' else {'text': self._get_picture_url(reg, reg.picture_data),
+                                                         'is_picture': True}
                        for column_name in column_names]
             return {'checked_in': self._is_checkin_visible(reg), 'columns': columns}
 
@@ -161,7 +161,8 @@ class RHParticipantList(RHRegistrationFormDisplayBase):
         is_participant = self.event.is_user_registered(session.user)
         registrations = sorted(_deduplicate_reg_data(_process_registration(reg, column_names)
                                                      for reg in query if reg.is_publishable(is_participant)),
-                               key=lambda reg: tuple(x['text'].lower() for x in reg['columns']))
+                               key=lambda reg: tuple(x['text'].lower() for x in reg['columns']
+                                                     if not x.get('is_picture')))
         return {'headers': headers,
                 'rows': registrations,
                 'show_checkin': any(registration['checked_in'] for registration in registrations),
@@ -174,7 +175,7 @@ class RHParticipantList(RHRegistrationFormDisplayBase):
             def _content(column_id, is_picture):
                 if column_id in data_by_field:
                     if is_picture:
-                        return self._picture_url(reg, data_by_field[column_id].field_data_id)
+                        return self._get_picture_url(reg, data_by_field[column_id])
                     return data_by_field[column_id].get_friendly_data(for_humans=True)
                 elif (column_id in active_fields and active_fields[column_id].personal_data_type is not None and
                         active_fields[column_id].personal_data_type.column is not None):
@@ -595,18 +596,18 @@ class RHParticipantListPictureDownload(RHParticipantList):
         RHParticipantList._check_access(self)
         if registration_settings.get(self.event, 'merge_registration_forms'):
             if 'picture' not in registration_settings.get(self.event, 'participant_list_columns'):
-                raise Forbidden
+                raise Forbidden('Picture column disabled')
             if not self.data.field_data.field.personal_data_type:
                 # only main picture from personal data is in merged form
-                raise Forbidden
+                raise Forbidden('Picture field is not the standard one')
         else:
             participant_list_form_columns = registration_settings.get(self.event, 'participant_list_form_columns')
             if (self.data.field_data.field_id not in
                     participant_list_form_columns[str(request.view_args['reg_form_id'])]):
-                raise Forbidden
+                raise Forbidden('Picture field is not exposed in participant list')
         is_participant = self.registration.event.is_user_registered(session.user)
         if not self.registration.is_publishable(is_participant):
-            raise Forbidden
+            raise Forbidden('Participant is not published')
 
     def _process_args(self):
         RHParticipantList._process_args(self)
@@ -622,13 +623,15 @@ class RHParticipantListPictureDownload(RHParticipantList):
                      .join(RegistrationFormFieldData.field)
                      .join(RegistrationData.registration)
                      .join(Registration.registration_form)
-                     .options((joinedload('registration').load_only(Registration.id,
-                                                                    Registration.consent_to_publish,
-                                                                    Registration.participant_hidden,
-                                                                    Registration.state,
-                                                                    Registration.is_deleted))
+                     .options(joinedload('registration').load_only(Registration.id,
+                                                                   Registration.event_id,
+                                                                   Registration.consent_to_publish,
+                                                                   Registration.participant_hidden,
+                                                                   Registration.state,
+                                                                   Registration.is_deleted)
                               .joinedload('registration_form')
                               .load_only(RegistrationForm.id,
+                                         RegistrationForm.event_id,
                                          RegistrationForm.publish_registrations_participants,
                                          RegistrationForm.publish_registrations_public,
                                          RegistrationForm.publish_registrations_duration),
@@ -637,4 +640,6 @@ class RHParticipantListPictureDownload(RHParticipantList):
         self.registration = self.data.registration
 
     def _process(self):
+        if not self.data or self.data.storage_file_id is None:
+            raise NotFound('Participant has no picture')
         return self.data.send()